requirements.txt
Flask==2.2.2
cassandra-driver==3.25.0
xlrd==2.0.1
numpy==1.21.6
pandas==1.1.5
seaborn==0.12.0
pymongo
sklearn